Move types (#1812)
* init

* add pass config type
